+++ Robust.Client/UserInterface/CustomControls/Tooltip.cs
@@ -1,4 +1,5 @@
-﻿using Robust.Client.UserInterface.Controls;
+using Robust.Client.UserInterface.Controls;
+using Robust.Shared.Timing;
 
 namespace Robust.Client.UserInterface.CustomControls
 {
@@ -12,6 +13,11 @@ namespace Robust.Client.UserInterface.CustomControls
             set => _label.Text = value;
         }
 
+        /// <summary>
+        /// Should we track the mouse cursor.
+        /// </summary>
+        public bool Tracking = false;
+
         public Tooltip()
         {
             var vbox = new BoxContainer()
@@ -23,6 +29,16 @@ namespace Robust.Client.UserInterface.CustomControls
             AddChild(vbox);
 
             vbox.AddChild(_label = new Label());
+        }
+
+        protected override void FrameUpdate(FrameEventArgs args)
+        {
+            base.FrameUpdate(args);
+
+            if (!Tracking)
+                return;
+
+            Tooltips.PositionTooltip(this);
         }
     }
 }